Fix a hang when an error occurs in a transaction block with replication mode
When an error occurs at a backend node in a transaction block with
replication mode, pgpool-II sends an invalid query to other
backends to sync transaction states. However, previously, this was
not sent to master node, and pgpool-II was waiting for master's
response forever.

This bug is caused to a specification change since 3.3.0 that load
balance is allowed in a transaction block with replication mode.

To fix this, pgpool sends the invalid query to all backends,
including master, except the node which the original query was sent.