Fixing a problem in elog.c when forwarding a message to frontend clients.
Error while forwarding the message to the frontend client can potentially cause
the infinite recursion and terminate the child with ERRORDATA_STACK_SIZE exceeded error.
Fix is to check the recursion depth of send_message_to_frontend() function
and pulling the plug if it is more than two levels deep.