Save pgpool_status to disk each time when it is changed.
Currently pgpool-II does not save pgpool_status file when it's
changed. Instead it saves when pgpool-II explicitly is going
shutdown. This is fine in most cases and it saves slight CPU cycles.
But what if pgpool-II killed abnormally? Or the system forces
pgpool-II to go down by OOM killer?  In these cases next pgpool-II
restarting could load obsolete status file.

This fix is a guard against such that situation by saving the
pgpool_status file each time when it's changed.