Add infrastructure to properly manage Parse/Bind/Close responses.
For this purpose, an ordered list to record Parse/Bind/Close message
added to session context. Functions to manage the list are also
added. Before there's a global variable to manage this, but
unfortunately now messages waiting for response it not only one. This
gives a confusion, for example CloseComplete() removes wrong message
data.