The next brick in the watchdog enhancement project, The majority of this commit contains the integration of lifecheck process with the new watchdog, Currently only lifecheck using heartbeat is working and the query mode still has few TODOs.
The new design of watchdog decouples the life checking from the watchdog, Now
the lifecheck process communicates with the watchdog using the unix domain IPC
sockets and forward the events to watchdog using JSON data,
This effectively means after this we can easily use any external heartbeat or
life-check system with the pgpool-II watchdog.

Following are the remaining TODOs..
Escalation to master pgpool-II and virtual-IP monitoring system is still half cooked.
Removal of all the traces of old watchdog
Integration  of interlocking watchdog commands
Code organization and cleanup