Allow to run pgpool in debug mode in regression tests.
For this purpose, -d option is added to regress.sh, which exports an
environment variable "PGPOOLDEBUG" being set to "true". Also -d option
is added to pgpool_setup. If this option is given or the environment
variable "PGPOOLDEBUG" is given and the value is "true", start.sh is
created with -d option added to start up pgpool daemon.