Add new parameter "serialize_accept".
This parameter defines whether to serialize accept() call for incoming
client connections.  Default is off, which means no serializing (same
behavior as pgpool-II 3.4 or before).  If this is off, the kernel
wakes up all of pgpool-II child process to execute accept() and one of
them actually accepts the incoming connection.  Problem here is,
because so my child process wake up at a same time, heavy context
switching occurred and the performance is affected.  This phenomena is
a classic problem called "the thundering herd problem".  By enabling
serialize_accept, only one of pgpool-II child process is woken up and
executes accept() and the problem can be avoided.

If child_life_time is enabled, serialize_accept has no effect.  If you
are care about pgpool-II process memory leak or whatever potential
issue, you could use child_max_connections instead.  This is purely an
implementation limitation and maybe we could eliminate it in the
future.