Fix in memory query in streaming replication mode and extended protocol case.
In streaming replication mode and extended query case, before
returning cached results, we need to read and forward any pending
responses such as parse complete and bind complete etc. For this
purpose we send flush message and retrieve the packets and forward to
frontend. The logic was there, but it was broken.