Fix hang in failover when serialize_accept is on.
This is due to semaphore for the locking is being left locked when
pgpool child process exits. To prevent the problem, give SEM_UNDO flag
to semop(). Also to clarify which status pgpool child are in, ps
status is shown as "wait for accept lock" when they are waiting for
lock.