Watchdog node goes into the WD_WAITING_FOR_QUORUM state whenever the quorum is not present or lost. Although it is a good guard against the split-brain syndrome, but there is a problem with this technique. pgpool-II commands which require a cluster wide synchronization like interlocking commands start getting failed when the node is in waiting for quorum, as these commands requires a central coordinator node for processing
The fix for this is to remove the WD_WAITING_FOR_QUORUM state and make sure that
the cluster always elects the master node even when the quorum is not present.
But the trick is not to execute the escalation commands on master node when the
quorum is missing and waits until the quorum is complete. This new design
ensures that even when because of network partitioning or some other issue the
cluster gets multiple master nodes (split-brain syndrome) the VIP conflict will
still not happen and that multiple master node would be harmless.