Another fix for [pgpool-hackers: 1255] pgpool-II should fail to start when add request is rejected
I tried to fix the issue by setting the POOL_EXIT_FATAL return code in the
add request reject fatal error message, but, unfortunately that was not enough
to completely fix the issue.  As for the errors occurred in watchdog before it
is fully initialized the pgpool-II main process keep stuck waiting for the
SIGUSR2 signal from watchdog, so the main process should also be looking for
SIGCHLD from watchdog along with SIGUSR2 while it is waiting for the watchdog
process to initialize.

The patch is contributed by Yugo Nagata <nagata@sraoss.co.jp>