commit a suicide as soon an the network becomes unreachable
Once the cable on the system is unplugged or when the node gets isolated from the
cluster there is every likelihood that the backend healthcheck of the isolated node
start reporting the backend node failure and the pgpool-II proceeds to perform
the failover for all attached backend nodes. Since the pgpool-II is yet not
smart enough to figure out it is because of the network failure of its own
system and the backend nodes are not actually at fault, but are working properly.

So now when the network gets back the backend status of the node will be different
and incorrect from the other pgpool-II nodes in the cluster. So the ideal solution
for the situation is to make the pgpool-II main process aware of the network black out
and when the network recovers the pgpool-II asks the watchdog to sync again the state of
all configured backend nodes from the master pgpool-II node. But to implement this lot
of time is required, So until that time we are just opting for the easiest solution here
which is to commit a suicide as soon an the network becomes unreachable