pcp_watchdof_info utility enhancement
pcp_watchdog_info only gives the information of a single watchdog node which
might not be enough in some certain situations. And as we are currently in the
process of enhancing watchdog system so pcp_watchdog_info utility is also
getting its share of enhancements. The commit adds a little more information
about the watchdog cluster state and nodes in the output of pcp_watchdog_info.