Fixing network interface monitoring mechanism of watchdog.
When the network cable is unplugged some Linux flavors un-assign all the IP
addresses assigned to that network on which the cable was unplugged while other
keeps the IP addresses and only un-sets the LOWER_UP bit in the if_flags.
The problem was we were only considering the first case where all IP addresses
gets removed when the network is unplugged, and the mechanism was failing on the
system where the flag bit indicates the network cable unplug.
The fix is to, along with the address change notifications, the watchdog now
also listens to link change notifications (RTM_NEWLINK and RTM_DELLINK) and upon
receiving the link change event it checks the ifa_flags bits to decide if the
link is active or inactive

The commit also adds a new configuration parameter "wd_monitoring_interfaces_list"
to make this network interface monitoring configurable. This new config parameter
can be assigned the list of interface names to monitor and watchdog will continue
if at least one interface is active from that list.