Properly process empty query (all comments)
Per bug#154. Pgpool-II recognizes empty query, for example "/* DBD::Pg
ping test v3.5.3 */" (note that no ';') as an error and emits message
" Unable to parse the query". This is because raw_parser() cannot
distinguish an empty query from an error query due to the poor API
design. To fix this new out parameter "*error" added to the
function. If an error occurred, the variable is set to true and caller
can know the reason why raw_parser() returns NIL is because of error
or empty query. See [pgpool-hackers: 1359] for additional information.