Fix pgpool hung after receiving error state from backend
Per bug #169.

This could happend if we execute an extended protocol query and it
fails. After an error is received the "ignore till sync flag" is set
and remained even if sync message was actually received. Thus any
subsequent query (in the case above "DEALLOCATE message") is not
procesed and pgpool waits for message from frontend and backend, and
pgpool stucks here because no message will arrive from both side.

To fix this, unconditionally reset the "ignore till sync flag" in
ReadyforQuery(). This is safe because apparently we already received
the ready for query message.