Fix pgpool hang bug (deal with bug 167)
pgpool-II 3.5 or after in streaming replication mode does not wait for
response from each phase such parse, bind anymore.  However, if
do_query is called, it sends flush message to retrieve the result of
system catalog look up. This is only sent to primary node which may
results in retrieving previous message results, for example parse
complete. If standby is assigned to load balance node, the node does
not return parse complete message, which will cause a problem in
bug167 case, because parse message for "BEGIN" was sent to both the
primary and the standby. Fix is, send flush message in do_query if the
load balance node is one of standbys.