Fix extended query hangs
Deal with bug 167. The problem is in the following sequence:

Parse(BEGIN)
Bind
:
Execute
Parse(SELECT)
Bind
:

(Notice that no Sync message between Execute and Parse)
BEGIN is sent to all db nodes while select is sent to one of db
nodes. So pgpool is confused which DB node to read the response. To
fix this, after each Execute, send flash message to backend to get
response. This will ensure that the responses of BEGIN etc. are read
from all DB nodes while the responses of SELECT etc. are read from
only one of DB node.