Fix for [pgpool-general: 4569] PGPoolII-3.5 : segfault
PostgreSQL's memory and exception manager APIs adopted by the pgpool 3.4 are not
thread safe and are causing the segmentation fault in the watchdog lifecheck
process, as it uses the threads to ping configured trusted hosts for checking
the upstream connections.
Fix is to remove threads and use the child process approach instead.