Fix extended protocol hang with empty query
The fixes related to extended protocol cases in 3.5.1 broke the case
of empty query.  In this case backend replies with "empty query
response" which is same meaning as a command complete message. Problem
is, when empty query response is received, pgpool does not reset the
query in progress flag thus keeps on waiting for backend. However,
backend will not send the ready for query message until it receives a
sync message. Fix is, resetting the in progress flag after receiving
the empty query response and reads from frontend expecting it sends a
sync message.

Per bug 190.