Streamlining the pgpool configuration system (Work in progress).
This commit builds the framework for managing the pgpool-II configuration
parameters. It is the first step towards implementing the session level
SET command functionality in pgpool to allow the users to alter the pgpool
config parameters for a particular session only and reset it back to the
original value after the user session disconnects.
The overall design of the configuration framework is inspired by the
PostgreSQL's GUC and some of the more visible changes introduced by the
framework are.

-- There is no more a long list of if-else statements in pool_get_config function
and now a simple loop can be used to alter/initialize/print all the pgpool-II
config variables.
-- The framework keeps track of values assigned to parameters in BOOT or INIT
contexts and later value of the config parameters can be reset to that value.

In addition to the framework, the commit also revisits the internal data types
of the configuration parameters and made the following changes
-- Changes the type of all on/off type config parameters from integer to boolean type
-- All the string (char*) type parameters that can only be assigned from the
particular set of values (e.g. master_slave_sub_mode can assign with either
"stream" or "Slony") are changed to Enum type parameters.