Fixing the crash caused by a mistake made by previous commit of config framework.
The commit contains the fixes for problems left out of previous commit, one of
that was causing the segmentation fault. Along with that the commit suppresses
the warning message emitted at the time of reloading the configuration file for
config parameters that are not allowed to be changed in reload config context.