Fix hang when portal suspend received
When portal suspend message is received, it's not enough to forward it
to the client.  Since backend expects to receive execute message,
trying to read further more messages from backend will never
succeed. To fix this, turn off the query in progress flag, which will
pole incoming message from the client.

Per bug 230.