Make authentication error message more user friendly.
When attempt to connect to backend (including health checking),
currently we emit something like:

"failed to authenticate"
"invalid authentication message response type, Expecting 'R' and received '%c'"

This is too cryptic for users. So emit the error message returned from
backend (for example, "sorry, too many clients already" instead of
this. By this, user could have cleaner idea on what's going on.

Since this is not a bug fix, committed in master branch only for now.