Fix for: [pgpool-general: 4997] Avoiding downtime when pgpool changes require a restart
To fix this, The verification mechanism of configuration parameter values is
reversed, previously the standby nodes used to verify their parameter values
against the respective values on the master pgpool-II node and when the
inconsistency was found the FATAL error was thrown, now with this commit the
verification responsibility is delegated to the master pgpool-II node.
Now the master node will verify the configuration parameter values of each
joining standby node against its local values and will produce
a WARNING message instead of an error in case of a difference.
This way the nodes having the different configurations will also be allowed to
join the watchdog cluster and the user has to manually look out for the
configuration inconsistency warnings in the master pgpool-II log to avoid the
surprises at the time of pgpool-II master switch over.